move identifer filters to identifiers module
