+++ tests/unit/memory/memory_interface/test_interface_attack_results.py
@@ -8,8 +8,8 @@ from typing import TYPE_CHECKING, Optional
 from pyrit.common.utils import to_sha256
 from pyrit.identifiers import ComponentIdentifier
 from pyrit.identifiers.atomic_attack_identifier import build_atomic_attack_identifier
+from pyrit.identifiers.identifier_filters import IdentifierFilter, IdentifierType
 from pyrit.memory import MemoryInterface
-from pyrit.memory.identifier_filters import IdentifierFilter, IdentifierType
 from pyrit.memory.memory_models import AttackResultEntry
 from pyrit.models import (
     AttackOutcome,
