+++ tests/unit/memory/test_identifier_filters.py
@@ -3,8 +3,8 @@
 
 import pytest
 
+from pyrit.identifiers.identifier_filters import IdentifierFilter, IdentifierType
 from pyrit.memory import MemoryInterface
-from pyrit.memory.identifier_filters import IdentifierFilter, IdentifierType
 from pyrit.memory.memory_models import AttackResultEntry
